wormhole-attester: port the v1.3.0 healthcheck hotfix (#454)
* wormhole-attester: port the v1.3.0 healthcheck hotfix

This PR fixes a missing assignment to HealthCheckState's enable
flag. This would cause the attester to ignore all settings (including
defaults) and keep the enable flag set to false because we would never
bother setting it from attestation_cfg that came at runtime.

The bug is fixed and the healthcheck behavior is hardened. Currently,
the lazy_static block will explicitly initialize the global with
default values from attestation_cfg (enable is true, window size is
100). This prevents a similar human error from disabling the
healthcheck. Secondly, main.rs overwrites the default values using the
HealthCheckState::new() constructor, instead of individual struct
member assignments. This ensures that all necessary values or their
defaults are passed as the desired healthcheck state.

* p2w_autoattest.py: Get debug output from first attestation

* wormhole-attester: fix ()-less function bug, harden usize convertion

This change fixes a bug that would cause a gigantic allocation to
happen, after the attestation_cfg::default_healthcheck_window_size
function is taken without being called and converted to a very large
usize value. This would result in a big allocation which would crash
the attester. At the core it's a misunderstanding about "Fn() as
usize" conversion which unfortunately is safe Rust.

* Trigger build

* wormhole-attester: Fix a misuse of a mutex guard

This commit changes the `let hc = <lock the mutex>` into a direct
assignment to the locked mutex guard. Rust mutices are typically
scope-based which makes using them in variables more error prone.

* Dockerfile.client: Remove unused ADD wormhole-attester.

This speeds up the build of the container, and prevents useless
triggers from the attester codebase

* p2w_autoattest.py: simplify log filtering to just info