Set the RPC url of solana-devnet bridge client image to localhost
The previous value - service address - caused the pod to never reach a ready state (because the client couldn't reach the RPC server), and that caused the server to never serve requests because one of the container in the pods behind the service (the client one) never reached a ready state. In other words, this removes a circular dependency between the pod and its service.