Don't exit multisig script when vault exists
It looks like this condition should have matched the `res.returncode == errno.EEXIST`, but the return code is coming back as `1`, probably due to the wrapping around the subprocess call.

So we match directly on the stdout of the subprocess call and prevent the script from exiting when the output says the vault exists.