wormhole_attester client: Change "not enough data" HTTP code to 503
Previously, we used 307 Internal Redirect to tell k8s that we're still
waiting for the healthcheck window to fill up. Unfortunately, k8s
assumes 3XX as a success, which wrongly tells it that the service is
ready. Changing to 503 causes the expected probe failure to appear
when the healthcheck state is not yet determined.