[hermes] Improve latest feeds rest api (#752)
* [hermes] Improve latest feeds rest api

This change adds verbose and binary option to latest_price_feeds
endpoint. Unfortunately it exposes many internal information which
required touching different components to expose batch_vaa specific
information. The code is now coupled to batch_vaa and we need to
refactor it when we add other proof types (and eventually remove it when
it get deprecated).

* Update hermes/src/network/rpc/rest.rs

Co-authored-by: Reisen <Reisen@users.noreply.github.com>

* Update hermes/src/network/rpc/rest.rs

Co-authored-by: Reisen <Reisen@users.noreply.github.com>

* Fix merge

---------

Co-authored-by: Reisen <Reisen@users.noreply.github.com>