chore(target_chains/sui): update sui and wormhole dependency
Developers on Sui need to pin the same Sui version as their
dependencies and it enforces everyone to use the same Sui version.

Pyth and Wormhole Sui versions have been old and this change updates
them to the latest testnet version (1.19.1) to make integrations with us
easier. Wormhole has undergone a similar change and thereforce its
version has changed as well.

No contract deployment is needed for this change as it only unblocks
downstream consumers for compiling their contracts.