[contract_manager] make solana rpcs configurable (#1646)
I'm hitting errors using the mainnet-beta endpoint when trying to deploy the entropy contract. This change makes the RPCs configurable in the functions so I can pass in a different RPC in the script.

The right thing to do here is probably to load the solana rpc urls from configuration files in the chains directory, but doesn't seem worth it to do that now.