chore(docker): fix & rewrite Node Dockerfiles
The new versions of these are designed to properly take advantage of pnpm and to
also minimize output size.

The builder image no longer is built on layers starting with just the
package.json files.  The reason for this is that with pnpm's content-addressed
store and the docker buildkit cache mounts, rebuilding the layers adds only
trivial overhead if dependencies haven't changed, and removing the Docker
layering to attempt to cache images with just node modules reduces complexity
significantly.

The runner image is now based on alpine and is drastically smaller.  We use
`pnpm export` (or, in the case of nextjs, we use standalone builds) to populate
the runner images with only the necessary packages and dependencies to further
reduce image sizes.