fix(target_chains/fuel): update js sdk to properly gen and export types
This change adds type generation to the build path to have them in the
resulting library and also fixes the issues around type exports by
updating the typescript and using `export type *`. The problem was that
the normal `export *` adds a similar export to the produced js file but
there is no module to be exported from in our case.