fix(hermes): tune readiness probe
This change modifies the not_behind readiness indicator to use the
most recent slot from Pythnet accumulator instead of taking the max
because we think that during forks validators might return an absurdly
large slot. It also makes it more verbose by returning the indicators
and the initial values for them so we can investigate better if the
issue persists.