fix(governance/xc_admin_cli): remove duplicates in pub buffer init
The JS Set is not smart to support custom equality operator and for
object it only considers the equal objects the same.

Activate and approve instructions are also changed because the sqds/mesh
sdk uses an old Anchor version that uses a deprecated getRecentBlockHash
RPC method. This method is removed in Agave v2 and devnet is fully on
Agave v2 now.