fix(target_chains/near): update dependencies and code to make tests work (#2100)
The old workspaces package in near that is used as a test suite couldn't download the required near-sandbox for it from an external url and it resulted in integration tests failing.

The new workspaces package is called near-workspaces and its near-sandbox has some rust version requirements to fully work, otherwise it panics with CompilationError(PrepareError(Deserialization)). The rust version 1.79.0 that its sandbox uses works but the latest stable rust 1.82.0 still fails. This assumption is not documented anywhere and I guessed it based on this old issue.

In the journey to make it work with @Riateche, we ended up updating the SDKs hoping to fix the problem and it's probably not needed but as the changes are good and will save us time in the future, i'll be keeping them.