feat(hermes-client): allow passing fetch options
This PR enables passing the `RequestInit` options for `fetch` to individual
`HermesClient` methods.

This PR also fixes merging `AbortControllers`, which was previously not done
correctly (but also no way to pass an `AbortController` was actually exposed via
the public API, so this is not a breaking change), and upgrades typescript to
the catalog version, which is required to support `AbortSignal.any`.