chore: (partially) resume pre-commit CI action (#2463)
Summary
Partially resume pre-commit CI. The problem is that a dependency of the pythnet_sdk has updated so that it now requires a newer rust version. Upgrading our rust version causes other problems however. The right fix to this is probably to add a Cargo.lock file (but I don't have one unfortunately).

Luckily the failure happens in the test package, so as a temporary stopgap turn off the test part of the clippy check.