Add Cargo.lock and renable pre-commit CI (#2465)
Summary
pre-commit in CI was broken because (1) the latest version of a dependency of pythnet SDK needs a higher rust compiler version, and (2) there wasn't a Cargo.lock file for pythnet_sdk.

Using Cargo.lock files for libraries is a bit controversial in the Rust community, as some argue that no Cargo.lock ensures that you keep your library up-to-date with the latest version of dependencies. However, this of course means you don't have repeatable builds, so CI and such are problems. In our case, I think CI is more important than keeping up to date with dependencies (as we're essentially the only people using the library.)

How has this been tested?
I ran the clippy command locally. Also, if pre-commit passes in CI, then it's good to go.