chore: separate type imports from value imports
This PR separates type imports from value imports, in compliance with a newly
enforced eslint rule from the new @cprussin configs.

This rule is important as it helps prevent cases where tree shaking is less
effective due to modules getting included in the runtime build which are in fact
only used for type imports.