Only build stuff that requires Vercel env to build on Vercel
For many of our UI applications, in order to build the package, a number of
environment variables must be present.

The values for these env vars are stored in Vercel, however not everyone who
contributes to this repository will have access to Vercel.

This PR modifies the turbo config to not build those packages on `turbo build`.
This is generally probably the right thing since we don't want those packages to
build in Github Actions either, given the environment is not going to be present
in GHA and so the builds would likely fail there too.

After this PR, to build any packages that rely on Vercel environments, you will
run `turbo build:vercelonly`.  Running `turbo build` will build everything else
and will exclude anything that depends on the Vercel env.