feat(fortuna): refactor explorer endpoint & history query methods
This commit implements a few fixes for the fortuna explorer endpoint, and in
doing so I refactored the query methods.  We are now using a query builder
approach which enables the filters, limits, offsets, etc to all work correctly
regardless of the query supplied.

This commit also implements a new query param to filter on state.