fix(component-library): upgrade react-aria and react-aria-components
Our prior version of react-aria was susceptible to the bug at
https://github.com/adobe/react-spectrum/issues/6399, which is a race condition
that causes the table to sometimes fail to render items when rerendering.  I
didn't think our version was susceptible to this bug as I haven't personally
encountered it, but Tejas mentioned he saw it in the entropy explorer.  I expect
that upgrading should resolve the bug for good.