fix(entropy-explorer): fix infinite spinner
The `Paginator` suspense boundary was incorrectly referring to the `<Paginator>`
component in the fallback case, which was a circular reference and caused the
stack to blow up, caused the browser loading spinner to go on indefinitely, and
is causing a bunch of errors in the Vercel logs.  This commit fixes that issue.