fix(entropy): Only mark callbacks as failed if enough gas was provided, regardless of revert code (#3106)
Summary
This PR removes the check that allows us to mark callbacks as failed (even if enough gas wasn't provided) if they revert with a non-empty return code.

Rationale
When we wrote this initally, we thought that all out-of-gas reverts from calling contracts would bubble up and be caught by the ret.length == 0 condition. However, we have found that some contracts catch out-of-gas reverts and then revert with a different error code. This means we can end up calling these callbacks without enough gas.

Simple fix here is to require enough gas to be provided regardless of the revert code.