fix(insights): use web3.js Connection instead of PythConnection
Upon investigation, it seems the `PythConnection` from `@pythnetwork/client` is
extremely inefficient for a few reasons:

1. `PythConnection` requires loading _all_ account info when creating the
connection class, so that it can build a mapping from price account to program
account.  However IH never used this mapping in practice, so all of that was
wasted work.  This caused major issues for page load performance as loading all
account info from pythnet was a ton of parsing which locked up the browser for
multiple seconds.

2. `PythConnection` did not expose a mechanism to remove unused subscriptions

In doing this I also removed all the live prices contexts since none of that is
needed any more, as well as the call to initialize prices with the last
available price as that too looks unnecessary and redundant.

This change should _drastically_ improve performance in IH, especially during
page load.